requirements.txt
python-dotenv
praw
openai
pandas
requests
